[TrueCloudLab#3] neofsid: Rename contract to frostfsid
